remove-artifact: remove

+++ cli/cli/scripts/completions.sh
@@ -1,6 +1,11 @@
 #!/usr/bin/env bash
 
 set -e
+if [ -z "${CLI_BINARY_FILENAME}" ]; then
+    echo "Environment variable CLI_BINARY_FILENAME must not be empty" >&2
+    exit 1
+fi
+
 rm -rf completions
 mkdir completions
 
@@ -9,3 +14,5 @@ go build -o "${CLI_BINARY_FILENAME}" main.go
 for sh in bash zsh fish; do
 	"./${CLI_BINARY_FILENAME}" completion "$sh" > "completions/kurtosis.$sh"
 done
+
+rm "${CLI_BINARY_FILENAME}"
